requirements/base.txt
asgiref==3.7.2
asttokens==2.4.1
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
cryptography==42.0.5
decorator==5.1.1
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-stub==0.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
executing==2.0.1
factory-boy==3.3.0
Faker==23.2.1
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
microslac==0.0.17
microslac-core==0.0.0.14
moto==5.0.4
nodeenv==1.8.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pika==1.3.2
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py-partiql-parser==0.5.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.1
pytest-django==4.8.0
pytest-env==1.1.3
pytest-mock==3.14.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
s3transfer==0.10.1
shortuuid==1.0.13
six==1.16.0
split-settings==1.0.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
xmltodict==0.13.0
